Automated runs and evaluations of ecological niche models

Runs ecological niche models over all combinations of user-defined settings (i.e., tuning), performs cross validation to evaluate models, and returns data tables to aid in selection of optimal model settings that balance goodness-of-fit and model complexity. Also has functions to partition data spatially (or not) for cross validation, to plot multiple visualizations of results, to run null models to estimate significance and effect sizes of performance metrics, and to calculate niche overlap between model predictions, among others. The package was originally built for Maxent models (Phillips et al. 2006, Phillips et al. 2017), but the current version allows possible extensions for any modeling algorithm. The extensive vignette, which guides users through most package functionality but unfortunately has a file size too big for CRAN, can be found on the package website: <https://jamiemkass.github.io/ENMeval/articles/ENMeval-2.0.0-vignette.html>.

Details

See README for details.

Author

Maintainer: Jamie M. Kass jamie.m.kass@gmail.com

Authors:

Robert Muscarella

Peter J. Galante

Corentin Bohl

Gonzalo E. Buitrago-Pinilla

Robert P. Anderson

Other contributors:

Robert A. Boria [contributor]

Mariano Soley-Guardia [contributor]