ENMnulls statistics plot

Plot histogram of evaluation statistics for null ENM simulations

evalplot.nulls(
  e.null,
  stats,
  plot.type,
  facet.labels = NULL,
  metric.levels = NULL,
  return.tbl = FALSE
)

Arguments

e.null
ENMnull object

stats
character vector: metrics from results table to be plotted; examples are "auc.val" or "or.10p"; if more than one statistic is specified, the histogram plot will be faceted

plot.type
character: either "violin" or "histogram"

facet.labels
named list: custom names for the metric facets, in the form list(old_name = "new_name", ...)

metric.levels
character vector: custom factor levels for metrics; this controls the order that metric statistics are plotted

return.tbl
boolean: if TRUE, return the data frames of null results used to make the ggplot instead of the plot itself

Value

A ggplot of null model statistics.

Details

There are two variations for this plot, but both show null quantiles (0.01, 0.05, 0.5, 0.95, 0.99). For violin plots, the null distribution is displayed as a vertical shape (i.e., the violin) with horizontal lines showing the quantiles and the empirical value is plotted as a red point along the vertical axis. For histogram plots, the null distribution is displayed as a histogram with vertical lines showing the quantiles and the empirical value as a vertical red line on the distribution.